requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.0
openai==1.12.0
langchain-community==0.0.10
langchain-openai==0.0.2
streamlit==1.31.0
# streamlit-webrtc==0.44.3  # Optional - requires system dependencies (pkg-config, ffmpeg)
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
sendgrid==6.11.0
twilio==8.10.0
python-dateutil==2.8.2
websockets==12.0

